+++ app/src/scenes/settings/settings-modal.tsx
@@ -32,12 +32,31 @@ const SettingsModal = ({ navigation, visible, onClick }) => {
             }}
             icon={<Bell />}
           />
-          <SettingItem title="Personnaliser mes indicateurs" path="symptoms" navigation={navigation} onClick={onClick} icon={<Analytics />} />
-          <SettingItem title="Personnaliser mes objectifs" path="goals-settings" navigation={navigation} onClick={onClick} icon={<Goal />} />
+          <SettingItem
+            title="Personnaliser mes indicateurs"
+            path="symptoms"
+            navigation={navigation}
+            onClick={(path) => {
+              logEvents.logOpenIndicatorsSettings();
+              onClick(path);
+            }}
+            icon={<Analytics />}
+          />
+          <SettingItem
+            title="Personnaliser mes objectifs"
+            path="goals-settings"
+            navigation={navigation}
+            onClick={() => {
+              onClick();
+              logEvents.logOpenObjectivesSettings();
+            }}
+            icon={<Goal />}
+          />
           <SettingItem
             title="Saisir mon traitement"
             navigation={navigation}
             onClick={async () => {
+              logEvents.logOpenDrugSettings();
               onClick();
               const treatment = await localStorage.getMedicalTreatment();
               if (treatment) {