+++ CHANGELOG.md
@@ -1,3 +1,52 @@
+# [1.89.0](https://github.com/SocialGouv/jardinmental/compare/v1.88.1...v1.89.0) (2025-09-25)
+
+
+### Features
+
+* add EDIT_SURVEY and DELETE_INDICATOR event ([#493](https://github.com/SocialGouv/jardinmental/issues/493)) ([3fd570f](https://github.com/SocialGouv/jardinmental/commit/3fd570f1beb3c24129f63b0374aa8920bee78ddc))
+
+## [1.88.1](https://github.com/SocialGouv/jardinmental/compare/v1.88.0...v1.88.1) (2025-09-25)
+
+
+### Bug Fixes
+
+* add back today.svg and yesterday.svg import as there are still used ([#499](https://github.com/SocialGouv/jardinmental/issues/499)) ([9504422](https://github.com/SocialGouv/jardinmental/commit/950442238c64fdc242de058b9745d104fc0a295e))
+
+# [1.88.0](https://github.com/SocialGouv/jardinmental/compare/v1.87.0...v1.88.0) (2025-09-25)
+
+
+### Features
+
+* add new support page ([#494](https://github.com/SocialGouv/jardinmental/issues/494)) ([99979cb](https://github.com/SocialGouv/jardinmental/commit/99979cb90899a818a2ab33532c4ddfc852a2e961))
+
+# [1.87.0](https://github.com/SocialGouv/jardinmental/compare/v1.86.0...v1.87.0) (2025-09-25)
+
+
+### Features
+
+* add checklist logs ([#496](https://github.com/SocialGouv/jardinmental/issues/496)) ([ca355ab](https://github.com/SocialGouv/jardinmental/commit/ca355ab1ae15b76c7771b7aa143d9c018e7ca420))
+
+# [1.86.0](https://github.com/SocialGouv/jardinmental/compare/v1.85.0...v1.86.0) (2025-09-25)
+
+
+### Features
+
+* add indicator logs ([#497](https://github.com/SocialGouv/jardinmental/issues/497)) ([a188f17](https://github.com/SocialGouv/jardinmental/commit/a188f179ed014315176fb7f4f676afcfeb6e4a92))
+
+# [1.85.0](https://github.com/SocialGouv/jardinmental/compare/v1.84.0...v1.85.0) (2025-09-25)
+
+
+### Features
+
+* add log NPS ([#495](https://github.com/SocialGouv/jardinmental/issues/495)) ([cd8fc8c](https://github.com/SocialGouv/jardinmental/commit/cd8fc8c63d65c92f6ae33cf00189f290a1461eaa))
+
+# [1.84.0](https://github.com/SocialGouv/jardinmental/compare/v1.83.0...v1.84.0) (2025-09-25)
+
+
+### Features
+
+* add drug loading function ([#498](https://github.com/SocialGouv/jardinmental/issues/498)) ([b4bd218](https://github.com/SocialGouv/jardinmental/commit/b4bd218f47f8fc5ed0fdc297f2b5965515e56d4d))
+
 # [1.83.0](https://github.com/SocialGouv/jardinmental/compare/v1.82.0...v1.83.0) (2025-09-23)